SimPe 0.43d (QA) ---------------------------------------------------------     + Changed the package Core, Resource Index is now stored as a Hashtree     + Added dockable OW     + Added transparent GroupBox, CheckBox Controls     + Added a simplified XPTaskBox Control     + Added an easy to use WizardControl

+++ trunk/fullsimpe/SimPe Toolbox/SkinWorkshop.cs
@@ -420,7 +420,7 @@ namespace SimPe.Plugin
 						oci.ObjectType = ObjectTypes.Outfit;
 						SimPe.PackedFiles.Wrapper.ObjFunctionSort ofs = new SimPe.PackedFiles.Wrapper.ObjFunctionSort(0);
 						ofs.InGeneral = true;
-						oci.ObjectFunctionSort = (short)ofs.Value;
+						oci.ObjectFunctionSort = (uint)ofs.Value;
 						oci.ObjectFileName = cpf.GetItem("name").StringValue;
 						oci.Useable = true;
 						oci.Class = SimPe.Cache.ObjectClass.Skin;
